PodAffinity to make it preferred that game servers group toghether
Utilised preferredDuringSchedulingIgnoredDuringExecution to make it
so that the scheduler preferred putting all the game servers
as close as possible to each other, thus making it far easier to
do scaling down within Kubernetes.

This also includes an update to latest version of Godep, as well
as moving client-go up to 3.0.0 beta, so we could take advantage
of the PodAffinity functionality.